requirement.txt
requests
beautifulsoup4
pandas
mysql-connector-python
Flask
